requirements.txt
numpy
openai
pynput
python-dotenv
scipy
sounddevice
SpeechRecognition
